refactor: delete board module provider
