+++ src/app.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+
 import { BoardsModule } from './boards/board.module';
-import { typeORMConfig } from './configs/typeorm.config';
+import { typeORMConfig } from './database/typeorm.config';
 
 @Module({
   imports: [TypeOrmModule.forRoot(typeORMConfig), BoardsModule],
