+++ src/boards/board.module.ts
@@ -1,12 +1,15 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+
 import { BoardsController } from './boards.controller';
 import { BoardRepository } from './board.repository';
 import { BoardsService } from './boards.service';
+import { TypeOrmExModule } from '../database/typeorm-ex.module';
+import { Board } from './board.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([BoardRepository])],
+  imports: [TypeOrmExModule.forCustomRepository([BoardRepository])],
   controllers: [BoardsController],
-  providers: [BoardsService, BoardRepository],
+  providers: [BoardsService],
 })
 export class BoardsModule {}
