+++ src/boards/board.repository.ts
@@ -1,8 +1,21 @@
-import { BoardNewRepository } from './typeorm.decorator';
+import { CustomRepository } from '../database/typeorm-ex.decorator';
 import { Repository } from 'typeorm';
 import { Board } from './board.entity';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { BoardStatus } from './board-status.enum';
 
-@BoardNewRepository(Board)
-export class BoardRepository extends Repository<Board> {}
+@CustomRepository(Board)
+export class BoardRepository extends Repository<Board> {
+  async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
+    const { title, description } = createBoardDto;
+    console.log(this);
+    const board = this.create({
+      title,
+      description,
+      status: BoardStatus.PUBLIC,
+    });
+    await this.save(board);
+
+    return board;
+  }
+}
