+++ src/database/typeorm-ex.module.ts
@@ -1,9 +1,9 @@
 import { DynamicModule, Provider } from '@nestjs/common';
 import { getDataSourceToken } from '@nestjs/typeorm';
 import { DataSource } from 'typeorm';
-import { TYPEORM_CUSTOM_REPOSITORY } from './typeorm.decorator';
+import { TYPEORM_CUSTOM_REPOSITORY } from './typeorm-ex.decorator';
 
-export class TypeOrmModule {
+export class TypeOrmExModule {
   public static forCustomRepository<T extends new (...args: any[]) => any>(repositories: T[]): DynamicModule {
     const providers: Provider[] = [];
 
@@ -23,10 +23,11 @@ export class TypeOrmModule {
         },
       });
     }
+    console.log(providers);
 
     return {
       exports: providers,
-      module: TypeOrmModule,
+      module: TypeOrmExModule,
       providers,
     };
   }
